requirements.txt
# Core PyTorch packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model libraries
timm>=0.9.0

# Data processing and scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=8.3.0
h5py>=3.7.0

# Image processing and augmentation
opencv-python>=4.5.0
albumentations>=1.3.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars and utilities
tqdm>=4.62.0
rich>=12.0.0

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0

# Jupyter and development tools
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Testing and code quality
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0

# Additional useful packages
imageio>=2.19.0
scipy>=1.7.0 